Couple of fixes and the initial workings of the new Login Window.
Some fixes such as the pagination icon images and the start work of the
the new Login Window. Switched the resolution to more rectangular shape
to have more space to create UI.

Input groups were created but still needed to be expandable upon to be
able to be used as a password field.